back-end/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-use-email-as-username==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==26.0.0
mysqlclient==2.2.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
